requirements-saas.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis==5.1.1
rq==1.16.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==43.0.3
structlog==24.4.0
httpx==0.27.2
jinja2==3.1.4
python-multipart==0.0.12
itsdangerous==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.32.0
limits==3.13.0
slowapi==0.1.9
anthropic>=0.40.0
pydantic[email]>=2.0.0
aiosqlite>=0.20.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
resend>=2.0.0
apscheduler>=3.10.0
